Start to add agent code
I'm still formulating exactly how this works, but my idea behind this is
that we run each service as a daemon in a container on either the server
or the agent hosts.

In this commit I've added the agent code. The Agent class will run once
and check the status of the Drone agent itself, before deciding whether
to enable instance protection on the EC2 instance.

The server will simply fetch some stats from the API and publish them to
Cloudwatch, but that requires passing in an API token to authenticate
with Drone.

Both require some work around passing in the required environment
variables and running each of the classes as a long-running daemon.