Remove AwsInfo class
As I was lying in my bed last night unable to sleep I suddenly realised
the AwsInfo class was redundant as it's own class since the information
it gathers (instance_id) and autoscaling group name is only used by the
agent class, so will not be used elsewhere.

Merge it into the main agent class.

This also rearranges tests slightly to not specifically test the methods
that call the AWS API, but to test the expected return of the run
method. If it does something then it should return true (whether enable
or disable), and if it does nothing then it should return false.