Ensure proper exit from error
It appears that Thor does not allow exiting with an error code that is
not 0, unless that is raised by Thor itself. There is some [description
here][1].

I wish to be able to capture errors and put them into Logger, so I have
added a `begin`/`rescue` block capturing all errors.

I've added a very small test to capture that the errors are handled, but
I think there is some further testing that I'd like to be able to add to
ensure the service behaves as we expect it to. I'd like to add tests to
capture the exceptions raised by AWS errors which is possible to stub,
but I need to figure out how best to write this.

For now I'm just adding the simple block and test. It is suggested that
you can replace the `exit_on_failure?` method, but this did not capture
the exception like I wished it to[2], and I had trouble implementing
this against the correct test.

[1]: https://github.com/erikhuda/thor/issues/244
[2]: https://github.com/erikhuda/thor/issues/244#issuecomment-11116989