Add required_workers metric
The idea behind this metric is to be able to create autoscaling actions
based upon a single metric.

  If the number is positive, add that number of instances.

  If the number is negative, remove that number of instances.

  If it is 0, do nothing.

This adds some logic into the thing that produces the custom metrics
itself, since Cloudwatch and autoscaling policies are extremely
inflexible, and we wish to do some more calculated scaling.

During the office hours, we wish to always have at least one worker
idle. If there is more than one idle worker, we should terminate all
idle workers but leave one spare.

If one or more pending jobs are discovered, create the same amount of
instances to cope with high scaling.

However, outside of office hours, we don't want instances hanging
around wasting precious cash monies.

If one or more pending jobs come into the queue, we should that amount of instances.

If there are one or more idle workers, we should terminate that amount
of idle workers.

If neither of these things are true, do nothing.