Update logic for Drone v1
There is a difference in the API between 0.8 and 1.x. Previously, the
server was able to report upon the number of idle workers since they
held a connection open between the agent and the server.

Agents now do long polling of an http endpoint instead, so the Drone
server is unable to say how many agents are connected.

This changes the logic to calculate the number of instances in Pending
or InService in an autoscaling group, and removing the number of jobs
that are running will leave us with the total number of workers
available.