Update instance protection logic
Rather than have agents poll themselves and enable protection, this
updates the logic so that the server itself compares the current workers
in an autoscaling group with the instances reported by the Drone API as
running jobs. This entirely depends on having Drone agents set their
node name as their instance ID.

It's actually a lot simpler than the original code. We can enable and
disable instance protection all at once which is a cleaner operation.

It may be worth seeing if agents can be booted with scale in protection
on by default.