Re-use API class
Rather than pass in the same options twice to the Metrics and
InstanceProtection classes, get the result of the API class once, and
pass the results into the classes. It's probably better to do this since
we should guarantee that we're running both tasks against the same
information to ensure we don't provide inconsistent results.