Try to make fewer calls to AWS API
Since both actions require calls to the AWS API, this commit attempts to
make the call beforehand, and pass the same result into the process.
Similar to passing in the API, this achieves a more consistent
experience, and means we're making fewer calls to the API, which can
sometimes refuse due to rate limiting.

This makes a single call, and then passes in the required details to
both classes without any further queries.

We then have up to 3 further API calls, 1 to enable protection, 1 to
disable and adding metrics.