Protect against worker explosion
When this was implemented previously, we could say for certain how many
idle workers we had because Drone was aware of this. We are now assuming
how many idle workers we have, and this could be dangerous if they do
not provision correctly and accept jobs.

To protect against this, we calculate idle workers by the number of
instances in total minus the number of running jobs. If both pending
jobs and idle workers are positive, then we can be relatively sure that
the workers are not being correctly assigned jobs. If this is the case,
refuse to make any further changes to the capacity of the autoscaling
group while we diagnose and fix the issue.